META-Files korrigiert für Verwendung ohne Rahmenwerk
git-svn-id: svn://ntoserver/Datenverteiler/VRZ3/trunk/de.bsvrz.iav.gllib@18972 40c0be50-5d29-0410-901d-fb59edb5cafc

+++ junit/de/bsvrz/iav/gllib/gllib/StuetzstelleTest.java
@@ -49,7 +49,7 @@ public class StuetzstelleTest {
 
 		s = new Stuetzstelle<Integer>(10, 50);
 		assertEquals(10L, s.getZeitstempel());
-		assertEquals(50, s.getWert());
+		assertEquals((Integer) 50, s.getWert());
 	}
 
 	/**
